Renaming component and styles for better clarity
